+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 CHANGELOG
 =========
 
+1.5.0
+-----
+
+* New: Add `RequestBeforeSendLoggerPlugin` for logging dispatch of Guzzle requests.
+* Enhancement: Add event context entry `MonologGuzzleLogAdapter`.
+
 1.4.0
 -----
 
-BC: `SingleLineFormatter` for Monolog is removed. Use Monolog's `LineFormatter@~1.8.0` instead.
-Enhancement: Default `ExtendLoggerServiceProvider` monolog formatter now includes microseconds.
+* BC: `SingleLineFormatter` for Monolog is removed. Use Monolog's `LineFormatter@~1.8.0` instead.
+* Enhancement: Default `ExtendLoggerServiceProvider` monolog formatter now includes microseconds.
 
 1.3.2
 -----